ShutdownThread: Add support to wait for oem subsystem shutdown
We now delay powering off the device until the list of oem subsystems
which need to to a graceful shutdown and done with their shutdown tasks.
The list of subsystems that are critical for shutdown(if any) should be
present as a space separated list in the below property:

vendor.peripheral.shutdown_critical_list.

The state of each subsystem should be represented in:

vendor.peripheral.subsysName.state

The value should be either ONLINE or OFFLINE. We will delay the shutdown
till either all critical subsystem states go to OFFLINE or a timeout occurs,
whichever is sooner.

Change-Id: I09354ffd69720b2c125a167f831da4aeff382320