base: introduce Fontservice(squash)
Pie Bringup by @ZeNiXxX

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Fonts: patched a few need files for boot

Change-Id: I4c0b65396371dc73ad9d05beacbdde72a7793b0c
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Dynamic font fix for FDE devices

must be delayed until after decryption

Change-Id: I337af1c07b98bbda2b85020229b2905f9ec7f7ff
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Add progress for font apply [1/2]

since this can take a few seconds its nice to give
user some feedback that things are going on.
so add a broadcast for change font. This is still
not exactly sent when font apply is read but gives
some approx information

Change-Id: I11e85f26dc82b7fd81328801f1d3629deda81a66
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

FontService: fix potential NPE and handle disabled packages properly

This was noticed when trying to use apps that can "freeze" (disable) other apps
and "unfreeze" (enable) them when needed.

On "unfreezing", it seems that the FontService will try to determine if
the new package is a font provider, while the app is still frozen and
will result in a NPE in FontService. This happened when I try to use the
Island app (with Greenify integration) on the latest build of OmniROM.

This patch simply returns false in `isPackageFontProvider()` when package is not found,
and falls through to the REMOVE case if the function is false, since packages that are
disabled cannot be used as a system-wide font provider anyway.

Change-Id: Idf1e44ae72a67bbd433959f30b6f96cb42243036

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

FontService: fix null object reference

E AndroidRuntime:
 java.lang.NullPointerException: Attempt to invoke virtual method
 'android.content.res.AssetManager android.content.Context.getAssets()' on a
 null object reference

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Set font for remaining keyguard items

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>